implemented thread lock for parallel computation of node transition probabilities
